interfaces on each level except for service
